pyproject.toml
[project]
name = "pinterest-crawler"
version = "0.5.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["requests", "pillow", "typer", ]
[project.scripts]
crawler = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]